Use n_proof_of_work_bits to u32.

+++ crates/stwo-circuits/src/stark_verifier/proof.rs
@@ -18,7 +18,7 @@ pub const N_TRACES: usize = 4;
 pub struct ProofConfig {
     // TODO(lior): Add a check on the total security bits of the protocol given parameters
     //   such as `n_proof_of_work_bits`, `fri.n_queries`, etc.
-    pub n_proof_of_work_bits: usize,
+    pub n_proof_of_work_bits: u32,
 
     // AIR structure.
     pub n_preprocessed_columns: usize,
@@ -92,7 +92,7 @@ impl ProofConfig {
         } = pcs_config;
 
         Self {
-            n_proof_of_work_bits: *pow_bits as usize,
+            n_proof_of_work_bits: *pow_bits,
             n_preprocessed_columns,
             n_trace_columns: trace_columns_per_component.iter().sum(),
             n_interaction_columns,
